[breaking?] Make components resolver store refs
This stops the TargetComponentsResolver from resolving references to
definitions eagerly on resolution.  While this may break parts of the
CSP generator for now, it seems like it's necessary to get the generator
to emit the appropriate names for reference components.

Added a new method find() to perform back-resolution of components to
their appropriate members of the target, which the generator will need
to adopt.