Strip back world definition
Align to Ana's suggestions.

Now, the world specifically only contains things that are directly
adjacent to or above a target.  We also introduce the ability for
collection-target worlds to contain
subcomponents (for now, they contain all
subcomponents).

We'll need to revisit this later on to exclude any actors that are not part of
a collection-target world.