Simplify and improve event resolution
This may break some things, especially since I removed a lot of support code
that was special-casing TargetActors.  The tests pass with some
alteration, so we'll look at what happens in the plugin.

The idea here is I want to push as much event information pooling as possible
into this one location so the csp-gen and other aspects don't have to handle
things specially.