Make target and actor passing explicit through resolvers
We now have a ResolveContext holding the target and actor list.  This mainly
exists to push up any target finding as far away from the point of use as
possible, making it easier to debug if it goes wrong.  It also lets us
get away with not having to do a lot of object
graph faking in tests.